Module google_searchconsole1::api

Structs

ApiDataRow
There is no detailed description.

ApiDimensionFilter
A filter test to be applied to each row in the data set, where a match can return the row. Filters are string comparisons, and values and dimension names are not case-sensitive. Individual filters are either AND'ed or OR'ed within their parent filter group, according to the group's group type. You do not need to group by a specified dimension to filter against it.

ApiDimensionFilterGroup
A set of dimension value filters to test against each row. Only rows that pass all filter groups will be returned. All results within a filter group are either AND'ed or OR'ed together, depending on the group type selected. All filter groups are AND'ed together.

BlockedResource
Blocked resource.

Image
Describe image data.

MobileFriendlyIssue
Mobile-friendly issue.

ResourceIssue
Information about a resource with issue.

RunMobileFriendlyTestRequest
Mobile-friendly test request.

RunMobileFriendlyTestResponse
Mobile-friendly test response, including mobile-friendly issues and resource issues.

SearchAnalyticsQueryRequest
There is no detailed description.

SearchAnalyticsQueryResponse
A list of rows, one per result, grouped by key. Metrics in each row are aggregated for all data grouped by that key either by page or property, as specified by the aggregation type parameter.

SearchConsole
Central instance to access all SearchConsole related resource activities

SearchanalyticMethods
A builder providing access to all methods supported on searchanalytic resources. It is not used directly, but through the SearchConsole hub.

SearchanalyticQueryCall
Query your data with filters and parameters that you define. Returns zero or more rows grouped by the row keys that you define. You must define a date range of one or more days. When date is one of the group by values, any days without data are omitted from the result list. If you need to know which days have data, issue a broad date range query grouped by date for any metric, and see which day rows are returned.

SiteAddCall
Adds a site to the set of the user's sites in Search Console.

SiteDeleteCall
Removes a site from the set of the user's Search Console sites.

SiteGetCall
Retrieves information about specific site.

SiteListCall
Lists the user's Search Console sites.

SiteMethods
A builder providing access to all methods supported on site resources. It is not used directly, but through the SearchConsole hub.

SitemapDeleteCall
Deletes a sitemap from this site.

SitemapGetCall
Retrieves information about a specific sitemap.

SitemapListCall
Lists the sitemaps-entries submitted for this site, or included in the sitemap index file (if sitemapIndex is specified in the request).

SitemapMethods
A builder providing access to all methods supported on sitemap resources. It is not used directly, but through the SearchConsole hub.

SitemapSubmitCall
Submits a sitemap for a site.

SitemapsListResponse
List of sitemaps.

SitesListResponse
List of sites with access level information.

TestStatus
Final state of the test, including error details if necessary.

UrlTestingToolMethods
A builder providing access to all methods supported on urlTestingTool resources. It is not used directly, but through the SearchConsole hub.

UrlTestingToolMobileFriendlyTestRunCall
Runs Mobile-Friendly Test for a given URL.

WmxSite
Contains permission level information about a Search Console site. For more information, see Permissions in Search Console.

WmxSitemap
Contains detailed information about a specific URL submitted as a sitemap.

WmxSitemapContent
Information about the various content types in the sitemap.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.